jyk/pdf_decl_pipeline_test/requirements.txt
openai>=1.0.0
Pillow>=10.0.0
pymupdf>=1.24.0
python-dotenv>=1.0.0

